Move saveJson to util file
